move `MastNode` in submodule
