+++ core/src/lib.rs
@@ -82,6 +82,7 @@ pub mod crypto {
     }
 }
 
+// TODOP: fix
 pub mod mast;
 pub use mast::{MastForest, MastNode, MastNodeId, MerkleTreeNode};
